refactor(multisig_create): remove `rentCollector` argument
This commit reverts the multisig_create interface to the one that's currently deployed to prod. We do this to ensure the old versions of our SDK continue working the same way after we deploy the Rent Reclamation and Global Config features to prod. For creating multisigs with rent_collector set, we provide 2 alternatives: `multisig_create_v2` or `multisig_set_rent_collector`